feat(WHI-557): requalify mainnet gas profile on a canonical Mantle fork
Relabels the mock GasProfileMeasure samples as foundry_mock (excluded from
qualification), adds a reproducible fork-replay harness
(remeasure_mainnet_gas_profile) that measures real gas cost for the 7 base
2-hop zero-crossing route classes via eth_call/eth_estimateGas + stateOverride
against a read-only Mantle RPC, and regenerates the pinned mainnet gas-profile
artifact from the merged sample set. Only 3 route classes (v2/v2, v2/v3,
v2/moe) reach the 10-sample qualification threshold on real fork-replay
evidence; the rest correctly remain Unsupported per the ticket's fail-closed
requirement, rather than being filled with mock data.

Updates tests/pipeline_wiring.rs's V3/Moe scenarios to use the now-approved
[V2,V3]/[V2,Moe] route shapes (the old pure [V3,V3]/[Moe,Moe] keys only
qualified before under the looser foundry_mock-counts-too logic this PR
closes), and derives pool_types from route_key instead of a separate scalar
so the two can't drift out of sync again.

Updates docs/guides/implementation-roadmap.md so WHI-547's mainnet-deployment
prerequisites include WHI-551/WHI-557, and WHI-556 is identified as an E2E
profile-loader prerequisite rather than a mainnet gate.

Confirmed the 3 examples/protocols/agni/v3_monitor_executor_service(_1559).rs
live-pool-state test failures are pre-existing on a clean origin/dev
checkout, unrelated to this change; already tracked as DI-19 in
docs/DEFERRED_ISSUES.md, now with a follow-up ticket (WHI-628).

+++ src/amms/moe/math/bin_helper.rs
@@ -134,7 +134,7 @@ pub fn get_composition_fees(
     let (amount_x, amount_y) = packed_uint128_math::decode(amounts_in);
     let encoded_out = get_amount_out_of_bin(bin_reserves + amounts_in, shares, total_supply + shares)?;
     let (received_x, received_y) = packed_uint128_math::decode(encoded_out);
-    let total_fee = pair_parameter_helper::get_total_fee(parameters, bin_step)? as u128;
+    let total_fee = pair_parameter_helper::get_total_fee(parameters, bin_step)?;
 
     if received_x > amount_x {
         let delta_y = amount_y - received_y;
@@ -179,7 +179,7 @@ pub fn get_amounts(
         safe128(uint256x256_math::mul_shift_round_up(U256::from(bin_reserve_out), price, SCALE_OFFSET as u8)?)?
     };
 
-    let total_fee = pair_parameter_helper::get_total_fee(parameters, bin_step)? as u128;
+    let total_fee = pair_parameter_helper::get_total_fee(parameters, bin_step)?;
     let max_fee = fee_helper::get_fee_amount(max_amount_in, total_fee)?;
     let max_amount_in_total = max_amount_in + max_fee;
